// src/app/project/models/addnewsourcerequest.ts
export class AddNewSourceRequest {
    stateId: number;
    sourceName: string;
    sourceNo: number;
    updateFrequency: string;
    updateFrequencyId: number;
    updateType: string;
    ownerId: number;
    active: number;
    sourceTypeId: number;
    hasPhotos: boolean;
    photoPath: string;
    processedAt: string;
}